+++ .prettierrc
@@ -7,6 +7,7 @@
   "trailingComma": "es5",
   "importOrder": [
     "^react(.*)",
+    "^@router",
     "^@util/(.*)$",
     "^@asset/(.*)$",
     "^@store/(.*)$",
