+++ src/components/Login/GoogleLogin.tsx
@@ -1,0 +1,5 @@
+function GoogleLogin() {
+  return <p>로그인 중...</p>;
+}
+
+export default GoogleLogin;
