+++ src/main.tsx
@@ -1,10 +1,25 @@
+import { ThemeProvider, createTheme } from '@mui/material';
+import { koKR } from '@mui/material/locale';
+
 import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
+import { RouterProvider } from 'react-router-dom';
 
-import App from './App.tsx';
+import { router } from '@router';
+
+const theme = createTheme(
+  {
+    typography: {
+      fontFamily: 'Noto Sans KR, sans-serif',
+    },
+  },
+  koKR
+);
 
 createRoot(document.getElementById('root')!).render(
   <StrictMode>
-    <App />
+    <ThemeProvider theme={theme}>
+      <RouterProvider router={router} />
+    </ThemeProvider>
   </StrictMode>
 );
